Resolved a few minor compilation errors
I know that the tor protocol has been changed to prevent bypassing
relays in relays in this way but I wanted to have a look at how it
used to work. I fixed a couple of small issues along the way to get it
to compile:

1. configure.ac        -  update to use latest macros
2. Makefile.ac         -  changed to reflect renaming of a boost lib
3. SocksConnecton.cpp  -  fixed localAddress var def
3. *.cpp	       -  fixed ambigous placeholders reference in the
boost the boost bind overloaded functions. I think I fixed an old
workaround for a boost placeholders namespace bug which is fixed in the
lastest libraries.... anyway it felt like drinking a mug of cement.